renamed Users.rb to Pools.rb for clarity
Pools are groups of servers on a single hostname.  Users is
deprecated because users already refers to users for the webui,
and a pool shares much of the same risk, e.g., if the worker dies
then the shared pool servers also lose STD{IN/OUT/ERR}.

A linux user is still used, but due to it's heavily structured
nature (tightly fitting regex), its importance as being seen as
a "user" is diminished.