improved checks for shutdown
current mechanism is:
once a request is received by mrmanager, Time.now.to_f is saved
as a global. It sends this number to HQ and expects the HQ to
return the number in plaintext.

1) webui sends request to HQ (currently any user can stop
   any manager)

2) HQ preps worker to shut down and provides pubkey to manager

3) Worker acknowledges request and sends to HQ Time.now.to_f
   of acknowledgement, encrypted with HQ pubkey

4) HQ decrypts Time and sends back to worker, confirming action

5) Worker receives plaintext and compares to saved time value.
   Upon match, shut down
 5a) If mismatch, clear saved time value and ignore all other
     confirmations until a new request (#2 occurs)

rsa_time either is nil or a float. the manager can be brute-forced
repeatedly, but will never be able to succeed because
any value sent will be translated to float, which will fail ===
followed by the resetting of the value.

It's possible if the manager sets the time value but HQ is dead,
that it will never receive confirmation & shut down, but at that
point, the intent was already there and attacking has only
a single guess from there on anyway.