webui direct-to-worker functionality restored
recent changes to permissions requires the full hierarchy to be
traversed from the root to the pool to the server.

In the event a mrmanger is not employed, this meant that the pool
ability offered through mrmanager was incomplete and servers
could not be made directly.

this additional route is a bit hacky, but for the time being
is cleanly implemented even if it makes the code a bit more
magical.